--- homeworks/varsanyidaniel/hw_08_streamlit/a.py
@@ -1,26 +1,0 @@
-import streamlit as st
-import pandas as pd
-import requests
-import plotly.express as px
-
-API_KEY=st.secrets["weatherapp111"]["api_key"]
-print(API_KEY)
-
-""""
-@st.cache_data(ttl=3600)
-def collect_data(city, API_KEY):
-    url=f'http://api.openweathermap.org/data/2.5/weather?q={city}&appid={API_KEY}&units=metric'
-    response=requests.get(url)
-    data=response.json()
-    if response.status_code == 200:
-        return data
-    else:
-        st.error(f"Failed to fetch data. Error: {response.status_code}")
-
-st.title("Robot Dreams Python - Weather Map & Data Visualization App")
-
-city=st.sidebar.text_input("Enter city name:")
-
-data = collect_data(city)
-"""
-# streamlit run homeworks/varsanyidaniel/hw_08_streamlit/a.py
